pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "claytools"
description = 'A collection of convenience tools to help me do the stuff I do faster and with less friction.'
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
keywords = []
authors = [
  { name = "Alex Claydon", email = "alexclaydon@gmail.com" },
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: Implementation :: CPython",
  "Programming Language :: Python :: Implementation :: PyPy",
]
dependencies = [
  'black',
  'isort',
  'python-dotenv',
  'python-dateutil',
  'openai',
  'prompt-engine-py',
  'rich[jupyter]',
  'markdown',
  'mdx_wikilink_plus',
  'beautifulsoup4',
  'jupyterlab',
]
dynamic = ["version"]

[project.urls]
Documentation = "https://github.com/alexclaydon/claytools#readme"
Issues = "https://github.com/alexclaydon/claytools/issues"
Source = "https://github.com/alexclaydon/claytools"

[tool.hatch.version]
path = "__about__.py"

[tool.hatch.envs.default]
type = "virtual"
dependencies = [
  "pytest",
  "pytest-cov",
]

[tool.hatch.envs.default.scripts]
cov = "pytest --cov-report=term-missing --cov-config=pyproject.toml --cov=claytools --cov=tests {args}"
no-cov = "cov --no-cov {args}"

[[tool.hatch.envs.test.matrix]]
python = ["311"]

[tool.coverage.run]
branch = true
parallel = true
omit = [
  "claytools/__about__.py",
]

[tool.coverage.report]
exclude_lines = [
  "no cov",
  "if __name__ == .__main__.:",
  "if TYPE_CHECKING:",
]
